refactor: move the bertviz into explain button
